+++ src/pages/Login.js
@@ -1,4 +1,4 @@
-import LoginForm from '../components/Login/LoginForm';
+import LoginForm from '../components/Auth/LoginForm';
 
 export default function Login() {
   return <LoginForm></LoginForm>;
